requirements.txt
numpy
matplotlib
scipy
tensorflow-gpu>=2.0.0
opencv-python
